src/test/java/_0951_1000/_958_Check_Completeness_of_a_Binary_Tree/SolutionITest.java
package _0951_1000._958_Check_Completeness_of_a_Binary_Tree;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.MethodSource;
import utils.TreeNode;
import utils.TreeNodes;

import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.assertFalse;
import static org.junit.jupiter.api.Assertions.assertTrue;

/**
 * @author jacka
 * @version 1.0 on 12/9/2019
 */
class SolutionITest {

  @ParameterizedTest
  @MethodSource("solutionProvider")
  void testOnlineCase1(Solution solution) {
    final TreeNode root = TreeNodes.getTreeLevelOrder(1, 2, 3, 4, 5, 6);
    assertTrue(solution.isCompleteTree(root));
  }

  @ParameterizedTest
  @MethodSource("solutionProvider")
  void testOnlineCase2(Solution solution) {
    final TreeNode root = TreeNodes.getTreeLevelOrder(1, 2, 3, 4, 5, null, 7);
    assertFalse(solution.isCompleteTree(root));
  }

  @ParameterizedTest
  @MethodSource("solutionProvider")
  void testOnlineCase3(Solution solution) {
    final TreeNode root = TreeNodes.getTreeLevelOrder(1, 2, 3, 4, 5, 6);
    root.left.left.left = new TreeNode(8);
    assertFalse(solution.isCompleteTree(root));
  }


  static Stream<Solution> solutionProvider() {
    return Stream.of(
        new SolutionI()
    );
  }
}